+++ frontend/src/app/annotate/tableau-svg/tableau-svg.component.ts
@@ -1,151 +1,8 @@
-import { Component, ElementRef, Input, Output, ViewChild, EventEmitter, ChangeDetectorRef} from '@angular/core';
+import { Component, ChangeDetectorRef} from '@angular/core';
 import { CommonModule } from "@angular/common";
 import { Subject } from "rxjs";
-import { sum } from "@/util";
-
-interface Dimensions {
-    width: number;
-    height: number;
-}
-
-@Component({
-    selector: "[term]",
-    standalone: true,
-    imports: [],
-    templateUrl: "./tableau-svg-term.svg",
-})
-export class TableauTerm {
-    @Input()
-    public idx?: number;
-
-    @Input()
-    public label?: String;
-
-    /* background color, should probably be replaced by an enum type with the color lookup done elsewhere */
-    @Input()
-    public bg?: String;
-
-    @Input()
-    public term: String = '';
-
-    @Input()
-    public end: boolean = false;
-
-    @Input()
-    public rule?: String;
-
-    @ViewChild('idxText')
-    idxText?: ElementRef<SVGTextElement>;
-
-    @ViewChild('termText')
-    termText?: ElementRef<SVGTextElement>;
-
-    @ViewChild('labelText')
-    labelText?: ElementRef<SVGTextElement>;
-
-    @Output()
-    public onSize = new EventEmitter<Dimensions>();
-
-    padding = 5;
-    height = 20;
-
-    idxW = 0;
-    termX = 0;
-    termW = 0;
-    labelX = 0;
-    labelW = 0;
-    totalW = 0;
-
-    ngAfterViewChecked() {
-        this.idxW = this.idxText ? this.idxText.nativeElement.getComputedTextLength() + this.padding : 0;
-        this.labelW = this.labelText ? this.labelText.nativeElement.getComputedTextLength() + this.padding : 0;
-        this.termX = this.idxW;
-        this.termW = this.termText ? this.termText.nativeElement.getComputedTextLength() : 0;
-        this.labelX =  this.termX + this.termW + this.padding;
-        this.totalW = this.termText ? this.labelW + this.idxW + this.termText.nativeElement.getComputedTextLength() : 0;
-
-        this.onSize.emit({width: this.totalW, height: this.height});
-    }
-
-}
-
-@Component({
-    selector: "[tableau-tree]",
-    standalone: true,
-    imports: [TableauTerm],
-    templateUrl: "./tableau-svg-tree.svg",
-})
-export class TableauTree {
-    @Input()
-    tree: any;
-
-    levelHeight = 40;
-
-    /* Width of the tree node, has to be determined dynamically via onSize events from terms */
-    width = 0;
-    /* Height isn't stored in a member variable, because it can be computed as needed */
-
-    /* Dimensions of the biggest subtree.
-       Width is used to align all subtrees.
-       Height is used to determine the overall height of the tree. */
-    subWidth = 0;
-    subHeight = 0;
-
-    @Output()
-    public onSize = new EventEmitter<Dimensions>();
-
-    updateDimensions(size: Dimensions) {
-        this.width = Math.max(this.width, size.width!);
-        this.emitSize();
-    }
-
-    /* keep track of the current node's biggest subtree using onSize events */
-    updateSubDimensions(size: Dimensions) {
-        this.subWidth = Math.max(this.subWidth, size.width!);
-        this.subHeight = Math.max(this.subHeight, size.height!);
-        this.emitSize();
-    }
-
-    emitSize() {
-        this.onSize.emit({
-            width: Math.max(this.subWidth * (this.tree.subtrees?.length ?? 0), this.width),
-            height: this.subHeight + this.totalNodeHeight()
-        });
-    }
-
-    /* Determines the X coordinate of where subtree of index `idx` should be drawn */
-    subtreePosition(idx: number) {
-        let widthWithPadding = 1.15 * this.subWidth;
-        return widthWithPadding * idx - (widthWithPadding / 2) * (this.tree.subtrees.length - 1);
-    }
-
-    /* Generates a path definition for linking the end of the current node to subtree index `idx` */
-    subtreeLinkPath(idx: number) {
-        return [
-            // move to end of current node
-            `M 0 ${this.totalNodeHeight() - 15 }`,
-            // curve to half-way to subtree
-            `q 0 ${this.levelHeight/2} ${this.subtreePosition(idx)/2 } ${this.levelHeight/2}`,
-            // curve from half-way to subtree, to subtree position
-            `q ${this.subtreePosition(idx)/2} 0 ${this.subtreePosition(idx)/2} ${this.levelHeight/2}`
-        ].join(' ');
-    }
-
-    nodeHeight(node: any) {
-        // note that in this case height includes bottom padding for the node
-        return node.rule ? 60 : 40;
-    }
-
-    totalNodeHeight() {
-        return sum(this.tree.nodes.map(this.nodeHeight));
-    }
-
-    nodeY(idx: number) {
-        // y position of a given node is the sum of heights of all preceeding nodes
-        return sum(this.tree.nodes.slice(0, idx).map(this.nodeHeight));
-    }
-}
-
+import { Dimensions } from './types';
+import { TableauTree } from './tableau-tree.component';
 
 @Component({
     selector: "la-tableau-svg",
