+++ frontend/src/app/annotate/tableau-svg/tableau-tree.component.ts
@@ -1,0 +1,81 @@
+import { Component, Input, Output, EventEmitter } from '@angular/core';
+import { TableauTerm } from './tableau-term.component';
+import { Dimensions } from './types';
+import { sum } from "@/util";
+
+@Component({
+    selector: "[tableau-tree]",
+    standalone: true,
+    imports: [TableauTerm],
+    templateUrl: "./tableau-tree.component.svg",
+})
+export class TableauTree {
+    @Input()
+    tree: any;
+
+    levelHeight = 40;
+
+    /* Width of the tree node, has to be determined dynamically via onSize events from terms */
+    width = 0;
+    /* Height isn't stored in a member variable, because it can be computed as needed */
+
+    /* Dimensions of the biggest subtree.
+       Width is used to align all subtrees.
+       Height is used to determine the overall height of the tree. */
+    subWidth = 0;
+    subHeight = 0;
+
+    @Output()
+    public onSize = new EventEmitter<Dimensions>();
+
+    updateDimensions(size: Dimensions) {
+        this.width = Math.max(this.width, size.width!);
+        this.emitSize();
+    }
+
+    /* keep track of the current node's biggest subtree using onSize events */
+    updateSubDimensions(size: Dimensions) {
+        this.subWidth = Math.max(this.subWidth, size.width!);
+        this.subHeight = Math.max(this.subHeight, size.height!);
+        this.emitSize();
+    }
+
+    emitSize() {
+        this.onSize.emit({
+            width: Math.max(this.subWidth * (this.tree.subtrees?.length ?? 0), this.width),
+            height: this.subHeight + this.totalNodeHeight()
+        });
+    }
+
+    /* Determines the X coordinate of where subtree of index `idx` should be drawn */
+    subtreePosition(idx: number) {
+        let widthWithPadding = 1.15 * this.subWidth;
+        return widthWithPadding * idx - (widthWithPadding / 2) * (this.tree.subtrees.length - 1);
+    }
+
+    /* Generates a path definition for linking the end of the current node to subtree index `idx` */
+    subtreeLinkPath(idx: number) {
+        return [
+            // move to end of current node
+            `M 0 ${this.totalNodeHeight() - 15 }`,
+            // curve to half-way to subtree
+            `q 0 ${this.levelHeight/2} ${this.subtreePosition(idx)/2 } ${this.levelHeight/2}`,
+            // curve from half-way to subtree, to subtree position
+            `q ${this.subtreePosition(idx)/2} 0 ${this.subtreePosition(idx)/2} ${this.levelHeight/2}`
+        ].join(' ');
+    }
+
+    nodeHeight(node: any) {
+        // note that in this case height includes bottom padding for the node
+        return node.rule ? 60 : 40;
+    }
+
+    totalNodeHeight() {
+        return sum(this.tree.nodes.map(this.nodeHeight));
+    }
+
+    nodeY(idx: number) {
+        // y position of a given node is the sum of heights of all preceeding nodes
+        return sum(this.tree.nodes.slice(0, idx).map(this.nodeHeight));
+    }
+}
