+++ frontend/src/app/annotate/tableau-svg/types.ts
@@ -1,0 +1,5 @@
+
+export interface Dimensions {
+    width: number;
+    height: number;
+}
